requirements.txt
asyncio>=3.4.3
cryptography>=3.4.8
pycryptodome>=3.15.0
click>=8.0.0
colorama>=0.4.4
psutil>=5.8.0
requests>=2.25.1
pyyaml>=6.0
python-dateutil>=2.8.2
tabulate>=0.9.0
watchdog>=2.1.6
